Bug 323036 - Add support to other Localizable Resources
Since we added a new class in the hierarchy for LocalizationFile, we had to change ILocalizationSchema accordingly. Now we have: LocalizationFile (as upper class) and StringLocalizationFile, ImageLocalizationFile, SoundLocalizationFile, VideoLocalizationFile as derived classes.

Also we have implemented the LocalizationFileFactory who is capable of creating LocalizationFiles using reflection.

At this point, the build is broken since we still have to adapt lots of classes due to this change.

git-svn-id: file:///svnroot/tools/org.eclipse.sequoyah@2282 7995dc58-227d-0410-baee-59875a487a5b